+++ retdec/tools/decompiler.py
@@ -389,6 +389,13 @@ def parse_args(argv):
              'list of address ranges, e.g. 0x100-0x200,0x500-0x600).'
     )
     parser.add_argument(
+        '--decoding',
+        dest='sel_decomp_decoding',
+        metavar='TYPE',
+        choices=['everything', 'only'],
+        help='What should be decoded in a seletive decompilation? Default: everything'
+    )
+    parser.add_argument(
         '--no-addresses',
         dest='decomp_emit_addresses',
         action='store_false',
@@ -488,6 +495,7 @@ def main(argv=None):
     add_decompilation_param_when_given(args, params, 'decomp_unreach_funcs')
     add_decompilation_param_when_given(args, params, 'sel_decomp_funcs')
     add_decompilation_param_when_given(args, params, 'sel_decomp_ranges')
+    add_decompilation_param_when_given(args, params, 'sel_decomp_decoding')
     add_decompilation_param_when_given(args, params, 'decomp_emit_addresses')
     add_decompilation_param_when_given(args, params, 'generate_archive')
     decompilation = decompiler.start_decompilation(**params)